删除不必要的接口,提高手机存储大小的精确度
Former-commit-id: 47c9cb4c5034341b6860bb1b7b6ccab4a5a6cfce

+++ AutumnBox.Basic/Devices/DeviceInfoHelper.cs
@@ -139,7 +139,7 @@ namespace AutumnBox.Basic.Devices
                 return null;
             }
         }
-        public static int? GetStorageTotal(string id)
+        public static double? GetStorageTotal(string id)
         {
             return null;
         }
